+++ packages/common/src/index.ts
@@ -83,6 +83,7 @@ export * from './types/space';
 export * from './types/table';
 export * from './types/timeFrames';
 export * from './types/user';
+export * from './types/warehouse';
 export * from './utils/api';
 export { default as assertUnreachable } from './utils/assertUnreachable';
 export * from './utils/conditionalFormatting';